feat: move prune logic in separate file
